chore: move shared components
