+++ apps/masterbots.ai/app/u/[slug]/page.tsx
@@ -5,8 +5,8 @@ import {
 } from '@/services/hasura'
 import ThreadList from '@/components/shared/thread-list'
 import AccountDetails from '@/components/shared/account-details'
-import { BrowseCategoryTabs } from '@/components/routes/browse/browse-category-tabs'
-import { BrowseSearchInput } from '@/components/routes/browse/browse-search-input'
+import { CategoryTabs } from '@/components/shared/category-tabs/category-tabs'
+import { BrowseInput } from '@/components/shared/browse-input'
 
 export default async function BotThreadsPage({
   params
@@ -22,8 +22,8 @@ export default async function BotThreadsPage({
   })
   return (
     <div className="container">
-      <BrowseCategoryTabs categories={categories} />
-      <BrowseSearchInput />
+      <CategoryTabs categories={categories} />
+      <BrowseInput />
       <AccountDetails
         href={`/u/${params.slug}`}
         alt={user.username}
