+++ apps/masterbots.ai/components/shared/category-tabs/index.ts
@@ -1,0 +1,1 @@
+export * from './category-tabs'
